REFACTOR & STYLE : 기본 뷰 헤더 및 스타일 수정하자
