+++ src/component/erp/management/dashboard/index.js
@@ -1,0 +1,6 @@
+
+const DashboardComponent = (props) => {
+    
+}
+
+export default DashboardComponent;
