+++ src/component/erp/management/order/view-header-setting-modal-v3/SelectorField.view.js
@@ -26,8 +26,8 @@ export default function SelectorFieldView(props) {
                                     return (
                                         <MenuItem key={'order_header_title' + idx} value={data.id} sx={{ display: 'flex', padding: '5px 10px', justifyContent: 'space-around', alignItems: 'center' }}>
                                             <span>{data.headerTitle}</span>
-                                            <span> 
-                                                {props.erpDefaultHeader?.orderHeaderId === data.id ?
+                                            <span>
+                                                {props.defaultHeader?.orderHeaderId === data.id ?
                                                     <img
                                                         src='/assets/icon/bookmark_black_icon.png'
                                                         alt=""
