--- src/component/erp/management/release-complete/view-header-setting-modal-v4/DefaultTableField.view.js
@@ -1,26 +1,0 @@
-import { DefaultTableFieldWrapper } from "./ViewHeaderSettingModal.styled";
-
-export default function DefaultTableFieldView(props) {
-    return (
-        <DefaultTableFieldWrapper>
-            <div className='grid-wrapper'>
-                {props.defaultHeaderDetails.map(r => {
-                    let isChecked = props.isCheckedOne(r.matchedColumnName);
-                    return (
-                        <div
-                            key={r.matchedColumnName}
-                            className='grid-item'
-                            style={{
-                                background: isChecked && '#2c73d240',
-                                fontWeight: isChecked && 600
-                            }}
-                            onClick={() => props.onActionCheckHeaderDetail(r)}
-                        >
-                            {r.originCellName}
-                        </div>
-                    );
-                })}
-            </div>
-        </DefaultTableFieldWrapper >
-    );
-}
