+++ src/component/erp/management/sales/view-header-setting-modal-v3/ViewHeaderSettingModal.styled.js
@@ -94,6 +94,10 @@ const SelectorFieldWrapper = styled.div`
     .selector-box {
         width: 350px;
         padding: 10px;
+
+        @media screen and (max-width: 992px){
+            width: 100%;
+        }
     }
 
     .flex-box {
@@ -172,26 +176,6 @@ const InfoTextFieldWrapper = styled.div`
         padding: 0 10px;
     }
 
-    .input-item {
-        margin: 0 10px;
-        width: 330px;
-        height: 30px;
-        border: none;
-        border-bottom: 1px solid #d9d9d9;
-        border-radius: 3px;
-        padding: 0 15px;
-        font-size: 16px;
-        font-weight: 600;
-        box-sizing: border-box;
-    }
-
-    .view-header-title {
-        color: #000;
-        padding: 20px 10px;
-        display: flex;
-        align-items: center;
-    }
-
     .info-box{
         font-size: 14px;
         color: #2C73D2;
@@ -202,69 +186,37 @@ const InfoTextFieldWrapper = styled.div`
         }
     }
 
-`;
-
-const ViewHeaderInputFieldWrapper  = styled.div`
-    padding: 0 20px;
-
-    .input-item {
-        /* margin: 0 10px; */
-        width: 330px;
-        height: 30px;
-        border: none;
-        border-bottom: 1px solid #d9d9d9;
-        border-radius: 3px;
-        padding: 0 15px;
-        font-size: 16px;
-        font-weight: 600;
-        box-sizing: border-box;
-
-        @media screen and (max-width: 992px){
-            width: 100%;
-        }
-    }
-
-    .view-header-title {
-        color: #000;
-        display: flex;
-        align-items: baseline;
-        flex-wrap: wrap;
-    }
-
-    .info-box {
-        display: flex;
-        flex-wrap: wrap;
-        align-items: baseline;
-    }
-
     .image-button-el{
         position: relative;
         overflow: hidden;
-        width: 40px;
-        height: 40px;
+        width:30px;
+        height: 30px;
         background: white;
-        border: 1px solid #fff;
-        border-radius: 50%;
-        transition: 0.2s;
-        margin: 5px;
+        border: 1px solid #212529;
+        border-radius: 2px;
 
         cursor: pointer;
 
-        :hover {
-            transform: scale(1.05);
-            background-color: #e0e0e0;
+        &:hover{
+            transform: scale(1.02);
         }
+    }
+
+    .image-button-icon{
+        width:80%;
+        position: absolute;
+        top:50%;
+        left:50%;
+        transform: translate(-50%, -50%);
     }
 `;
 
-const TableOperatorFieldWrapper = styled.div`
+const ViewHeaderInputFieldWrapper  = styled.div`
+    padding: 10px 20px;
     display: flex;
-    flex-wrap: wrap;
-    gap: 10px;
-    margin-top: 10px;
-    padding: 0 20px;
     align-items: center;
     justify-content: space-between;
+    flex-wrap: wrap;
 
     .button-el{
         font-size: 14px;
@@ -295,6 +247,65 @@ const TableOperatorFieldWrapper = styled.div`
             font-size: 10px;
         }
     }
+
+    .input-item {
+        /* margin: 0 10px; */
+        width: 330px;
+        height: 40px;
+        border: none;
+        border-bottom: 1px solid #d9d9d9;
+        border-radius: 3px;
+        padding: 0 15px;
+        font-size: 14px;
+        font-weight: 600;
+        box-sizing: border-box;
+
+        @media screen and (max-width: 992px){
+            width: 100%;
+        }
+    }
+
+    .view-header-title {
+        color: #000;
+        display: flex;
+        align-items: center;
+        flex-wrap: wrap;
+    }
+
+    .info-box {
+        display: flex;
+        flex-wrap: wrap;
+        align-items: center;
+    }
+
+    .image-button-el{
+        position: relative;
+        overflow: hidden;
+        width: 40px;
+        height: 40px;
+        background: white;
+        border: 1px solid #fff;
+        border-radius: 50%;
+        transition: 0.2s;
+        margin: 5px;
+
+        cursor: pointer;
+
+        :hover {
+            transform: scale(1.05);
+            background-color: #e0e0e0;
+        }
+    }
+`;
+
+const TableOperatorFieldWrapper = styled.div`
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    margin-top: 10px;
+    padding: 0 20px;
+    align-items: center;
+    justify-content: space-between;
 
     @media all and (max-width: 992px){
         padding: 0 10px;
