--- src/component/erp/management/sales/view-header-setting-modal-v4/InfoTextField.view.js
@@ -1,11 +1,0 @@
-import { InfoTextFieldWrapper } from "./ViewHeaderSettingModal.styled";
-
-export default function InfoTextFieldView(props) {
-    return (
-        <InfoTextFieldWrapper>
-            <div className='info-box'>
-                {props.element}
-            </div>
-        </InfoTextFieldWrapper>
-    );
-}
