--- src/component/erp/management/sales/view-header-setting-modal-v4/TableOperatorField.view.js
@@ -1,9 +1,0 @@
-import { TableOperatorFieldWrapper } from "./ViewHeaderSettingModal.styled";
-
-export default function TableOperatorFieldView(props){
-    return (
-        <TableOperatorFieldWrapper>
-            {props.element}
-        </TableOperatorFieldWrapper>
-    );
-}
